extract suffix renderer to own file
